Fix query station id for hafas-client v5
